+++ tests/test_datasets/test_dataset.py
@@ -430,6 +430,54 @@ def test_create_ts_with_exog_int_timestamp():
     pd.testing.assert_frame_equal(ts.to_pandas(), expected_merged)
 
 
+@pytest.mark.parametrize(
+    "df_exog",
+    (
+        generate_ar_df(start_time="2023-01-01", periods=30, n_segments=1, freq="D").rename(
+            {"target": "target_exog"}, axis=1
+        ),
+    ),
+)
+@pytest.mark.parametrize("df", (generate_ar_df(start_time="2023-01-01", periods=30, n_segments=2, freq="D"),))
+def test_create_ts_with_exog_error(df, df_exog):
+    with pytest.raises(ValueError, match="There is a mismatch in feature sets between segments"):
+        _ = TSDataset(df=df, df_exog=df_exog, freq="D")
+
+
+def test_create_ts_different_feat_sets_df_exog_error():
+    df = generate_ar_df(start_time="2023-01-01", periods=30, n_segments=2, freq="D")
+
+    df_exog = pd.merge(
+        generate_ar_df(start_time="2023-01-01", periods=30, n_segments=1, freq="D").rename(
+            {"target": "target_exog1"}, axis=1
+        ),
+        generate_ar_df(start_time="2023-01-01", periods=30, n_segments=2, freq="D").rename(
+            {"target": "target_exog"}, axis=1
+        ),
+        on=["timestamp", "segment"],
+        how="right",
+    )
+    ts_df_exog = TSDataset.to_dataset(df_exog).dropna(axis=1)
+
+    with pytest.raises(ValueError, match="Given wide dataframe doesn't have all combinations of pairs"):
+        _ = TSDataset(df=df, df_exog=ts_df_exog, freq="D")
+
+
+def test_create_ts_different_feat_sets_df():
+    df = pd.merge(
+        generate_ar_df(start_time="2023-01-01", periods=30, n_segments=1, freq="D").rename(
+            {"target": "target_exog1"}, axis=1
+        ),
+        generate_ar_df(start_time="2023-01-01", periods=30, n_segments=2, freq="D"),
+        on=["timestamp", "segment"],
+        how="right",
+    )
+    ts_df = TSDataset.to_dataset(df).dropna(axis=1)
+
+    with pytest.raises(ValueError, match="Given wide dataframe doesn't have all combinations of pairs"):
+        _ = TSDataset(df=ts_df, freq="D")
+
+
 @pytest.mark.filterwarnings(
     "ignore: Timestamp contains numeric values, and given freq is D. Timestamp will be converted to datetime.",
     "ignore: You probably set wrong freq. Discovered freq in you data is N, you set D",
@@ -1017,45 +1065,6 @@ def test_to_dataset_on_integer_timestamp():
     assert pd.api.types.is_integer_dtype(df.index.dtype)
 
 
-def test_size_with_diff_number_of_features():
-    df_temp = generate_ar_df(start_time="2023-01-01", periods=30, n_segments=2, freq="D")
-    df_exog_temp = generate_ar_df(start_time="2023-01-01", periods=30, n_segments=1, freq="D")
-    df_exog_temp = df_exog_temp.rename({"target": "target_exog"}, axis=1)
-    ts_temp = TSDataset(df=TSDataset.to_dataset(df_temp), df_exog=TSDataset.to_dataset(df_exog_temp), freq="D")
-    assert ts_temp.size()[0] == len(df_exog_temp)
-    assert ts_temp.size()[1] == 2
-    assert ts_temp.size()[2] is None
-
-
-def test_size_target_only():
-    df_temp = generate_ar_df(start_time="2023-01-01", periods=40, n_segments=3, freq="D")
-    ts_temp = TSDataset(df=TSDataset.to_dataset(df_temp), freq="D")
-    assert ts_temp.size()[0] == len(df_temp) / 3
-    assert ts_temp.size()[1] == 3
-    assert ts_temp.size()[2] == 1
-
-
-def simple_test_size_():
-    df_temp = generate_ar_df(start_time="2023-01-01", periods=30, n_segments=2, freq="D")
-    df_exog_temp = generate_ar_df(start_time="2023-01-01", periods=30, n_segments=2, freq="D")
-    df_exog_temp = df_exog_temp.rename({"target": "target_exog"}, axis=1)
-    df_exog_temp["other_feature"] = 1
-    ts_temp = TSDataset(df=TSDataset.to_dataset(df_temp), df_exog=TSDataset.to_dataset(df_exog_temp), freq="D")
-    assert ts_temp.size()[0] == len(df_exog_temp) / 2
-    assert ts_temp.size()[1] == 2
-    assert ts_temp.size()[2] == 3
-
-
-def test_size_with_diff_number_of_features():
-    df_temp = generate_ar_df(start_time="2023-01-01", periods=30, n_segments=2, freq="D")
-    df_exog_temp = generate_ar_df(start_time="2023-01-01", periods=30, n_segments=1, freq="D")
-    df_exog_temp = df_exog_temp.rename({"target": "target_exog"}, axis=1)
-    ts_temp = TSDataset(df=TSDataset.to_dataset(df_temp), df_exog=TSDataset.to_dataset(df_exog_temp), freq="D")
-    assert ts_temp.size()[0] == len(df_exog_temp)
-    assert ts_temp.size()[1] == 2
-    assert ts_temp.size()[2] is None
-
-
 def test_size_target_only():
     df_temp = generate_ar_df(start_time="2023-01-01", periods=40, n_segments=3, freq="D")
     ts_temp = TSDataset(df=TSDataset.to_dataset(df_temp), freq="D")
@@ -1565,6 +1574,38 @@ def test_add_columns_from_pandas_update_regressors(
     assert sorted(ts.regressors) == sorted(expected_regressors)
 
 
+def test_add_columns_from_pandas_different_segment_sets_error_single(df_and_regressors, df_update_add_column):
+    df, df_exog, _ = df_and_regressors
+    ts = TSDataset(df=df, freq="D", df_exog=df_exog)
+
+    with pytest.raises(ValueError, match="There is a mismatch in segments between provided and expected sets"):
+        ts.add_columns_from_pandas(df_update=df_update_add_column.drop(columns=[("2", "new_column")]))
+
+
+def test_add_columns_from_pandas_different_segment_sets_error_multiple(df_and_regressors, df_update_add_column):
+    df, df_exog, _ = df_and_regressors
+    ts = TSDataset(df=df, freq="D", df_exog=df_exog)
+    df_update_add_column[("2", "new_column_1")] = 2
+    with pytest.raises(ValueError, match="There is a mismatch in feature sets between segments"):
+        ts.add_columns_from_pandas(df_update=df_update_add_column)
+
+
+def test_update_columns_from_pandas_different_segment_sets_error_single(df_and_regressors, df_update_update_column):
+    df, df_exog, _ = df_and_regressors
+    ts = TSDataset(df=df, freq="D", df_exog=df_exog)
+
+    with pytest.raises(ValueError, match="There is a mismatch in segments between provided and expected sets"):
+        ts.update_columns_from_pandas(df_update=df_update_update_column.drop(columns=[("2", "target")]))
+
+
+def test_update_columns_from_pandas_different_segment_sets_error_multiple(df_and_regressors, df_update_update_column):
+    df, df_exog, _ = df_and_regressors
+    ts = TSDataset(df=df, freq="D", df_exog=df_exog)
+    df_update_update_column[("2", "regressor_1")] = 2
+    with pytest.raises(ValueError, match="There is a mismatch in feature sets between segments"):
+        ts.update_columns_from_pandas(df_update=df_update_update_column)
+
+
 @pytest.mark.parametrize("update_slice", (slice(4, -4), slice(None, None, 2)))
 def test_update_columns_from_pandas_invalid_timestamps(df_and_regressors, update_slice, df_update_update_column):
     df, _, _ = df_and_regressors
@@ -1724,7 +1765,7 @@ def test_add_target_components_throw_error_inconsistent_components_names(
     ts_without_target_components, inconsistent_target_components_names_fixture, request
 ):
     inconsistent_target_components_names_df = request.getfixturevalue(inconsistent_target_components_names_fixture)
-    with pytest.raises(ValueError, match="Set of target components differs between segments '1' and '2'!"):
+    with pytest.raises(ValueError, match="Set of target components differs between segments!"):
         ts_without_target_components.add_target_components(target_components_df=inconsistent_target_components_names_df)
 
 
@@ -1747,7 +1788,7 @@ def test_add_prediction_intervals_throw_error_inconsistent_components_names(
     inconsistent_prediction_intervals_names_df = request.getfixturevalue(
         inconsistent_prediction_intervals_names_fixture
     )
-    with pytest.raises(ValueError, match="Set of prediction intervals differs between segments '1' and '2'!"):
+    with pytest.raises(ValueError, match="Set of prediction intervals differs between segments!"):
         ts_without_target_components.add_prediction_intervals(
             prediction_intervals_df=inconsistent_prediction_intervals_names_df
         )